FlowMod complete with TTL

+++ os/net/ipv6/res-flow-mod.h
@@ -1,18 +1,11 @@
 
 #include "lib/list.h"
 typedef struct flow_s {
-	//uint8_t flowid;
 	uip_ipaddr_t ipv6src;
-	//uint8_t srcmask;
 	uip_ipaddr_t ipv6dst;
-	//uint8_t dstmask ;
-	//uint16_t srcport;
-	//uint16_t dstport;
-	//uint8_t ipproto;
 	uint8_t action;
 	uip_ipaddr_t nhipaddr;
-	//uint8_t txpwr;
-	uint16_t ttl;
+	int ttl;
 }flow_s;
 
 typedef struct flow_entry_s{
